26 - address comments from review, delete memory_state_reader.go
